+++ src/transport/transport.c
@@ -30,8 +30,7 @@ static inline int poll(struct pollfd *fds __attribute__((unused)),
 #endif
 
 LIBPLDM_ABI_TESTING
-pldm_requester_rc_t pldm_transport_poll(struct pldm_transport *transport,
-					int timeout)
+int pldm_transport_poll(struct pldm_transport *transport, int timeout)
 {
 	struct pollfd pollfd;
 	int rc = 0;
@@ -48,7 +47,8 @@ pldm_requester_rc_t pldm_transport_poll(struct pldm_transport *transport,
 		return PLDM_REQUESTER_POLL_FAIL;
 	}
 
-	return PLDM_REQUESTER_SUCCESS;
+	/* rc is 0 if poll(2) times out, or 1 if pollfd becomes active. */
+	return rc;
 }
 
 LIBPLDM_ABI_TESTING
@@ -183,8 +183,8 @@ pldm_transport_send_recv_msg(struct pldm_transport *transport, pldm_tid_t tid,
 		/* 0 <= `timeval_to_msec()` <= 4800, and 4800 < INT_MAX */
 		rc = pldm_transport_poll(transport,
 					 (int)(timeval_to_msec(&remaining)));
-		if (rc != PLDM_REQUESTER_SUCCESS) {
-			return rc;
+		if (rc <= 0) {
+			break;
 		}
 
 		rc = pldm_transport_recv_msg(transport, tid, pldm_resp_msg,